Revert "[PM-1033] refactor(wip): make `AcceptUserCommand`"
This reverts commit dc1319e7fa70c4844bbc70e0b01089b682ac2843.
