+++ src/Core/Services/Implementations/OrganizationService.cs
@@ -1213,6 +1213,154 @@ public class OrganizationService : IOrganizationService
         await _mailService.SendOrganizationInviteEmailAsync(organization.Name, orgUser, new ExpiringToken(token, now.AddDays(5)), organization.PlanType == PlanType.Free, initOrganization);
     }
 
+    public async Task<OrganizationUser> AcceptUserAsync(Guid organizationUserId, User user, string token,
+        IUserService userService)
+    {
+        var orgUser = await _organizationUserRepository.GetByIdAsync(organizationUserId);
+        if (orgUser == null)
+        {
+            throw new BadRequestException("User invalid.");
+        }
+
+        if (!CoreHelpers.UserInviteTokenIsValid(_dataProtector, token, user.Email, orgUser.Id, _globalSettings))
+        {
+            throw new BadRequestException("Invalid token.");
+        }
+
+        var existingOrgUserCount = await _organizationUserRepository.GetCountByOrganizationAsync(
+            orgUser.OrganizationId, user.Email, true);
+        if (existingOrgUserCount > 0)
+        {
+            if (orgUser.Status == OrganizationUserStatusType.Accepted)
+            {
+                throw new BadRequestException("Invitation already accepted. You will receive an email when your organization membership is confirmed.");
+            }
+            throw new BadRequestException("You are already part of this organization.");
+        }
+
+        if (string.IsNullOrWhiteSpace(orgUser.Email) ||
+            !orgUser.Email.Equals(user.Email, StringComparison.InvariantCultureIgnoreCase))
+        {
+            throw new BadRequestException("User email does not match invite.");
+        }
+
+        return await AcceptUserAsync(orgUser, user, userService);
+    }
+
+    public async Task<OrganizationUser> AcceptUserAsync(string orgIdentifier, User user, IUserService userService)
+    {
+        var org = await _organizationRepository.GetByIdentifierAsync(orgIdentifier);
+        if (org == null)
+        {
+            throw new BadRequestException("Organization invalid.");
+        }
+
+        var usersOrgs = await _organizationUserRepository.GetManyByUserAsync(user.Id);
+        var orgUser = usersOrgs.FirstOrDefault(u => u.OrganizationId == org.Id);
+        if (orgUser == null)
+        {
+            throw new BadRequestException("User not found within organization.");
+        }
+
+        return await AcceptUserAsync(orgUser, user, userService);
+    }
+
+    public async Task<OrganizationUser> AcceptUserAsync(Guid organizationId, User user, IUserService userService)
+    {
+        var org = await _organizationRepository.GetByIdAsync(organizationId);
+        if (org == null)
+        {
+            throw new BadRequestException("Organization invalid.");
+        }
+
+        var usersOrgs = await _organizationUserRepository.GetManyByUserAsync(user.Id);
+        var orgUser = usersOrgs.FirstOrDefault(u => u.OrganizationId == org.Id);
+        if (orgUser == null)
+        {
+            throw new BadRequestException("User not found within organization.");
+        }
+
+        return await AcceptUserAsync(orgUser, user, userService);
+    }
+
+    private async Task<OrganizationUser> AcceptUserAsync(OrganizationUser orgUser, User user,
+        IUserService userService)
+    {
+        if (orgUser.Status == OrganizationUserStatusType.Revoked)
+        {
+            throw new BadRequestException("Your organization access has been revoked.");
+        }
+
+        if (orgUser.Status != OrganizationUserStatusType.Invited)
+        {
+            throw new BadRequestException("Already accepted.");
+        }
+
+        if (orgUser.Type == OrganizationUserType.Owner || orgUser.Type == OrganizationUserType.Admin)
+        {
+            var org = await GetOrgById(orgUser.OrganizationId);
+            if (org.PlanType == PlanType.Free)
+            {
+                var adminCount = await _organizationUserRepository.GetCountByFreeOrganizationAdminUserAsync(
+                    user.Id);
+                if (adminCount > 0)
+                {
+                    throw new BadRequestException("You can only be an admin of one free organization.");
+                }
+            }
+        }
+
+        // Enforce Single Organization Policy of organization user is trying to join
+        var allOrgUsers = await _organizationUserRepository.GetManyByUserAsync(user.Id);
+        var hasOtherOrgs = allOrgUsers.Any(ou => ou.OrganizationId != orgUser.OrganizationId);
+        var invitedSingleOrgPolicies = await _policyService.GetPoliciesApplicableToUserAsync(user.Id,
+            PolicyType.SingleOrg, OrganizationUserStatusType.Invited);
+
+        if (hasOtherOrgs && invitedSingleOrgPolicies.Any(p => p.OrganizationId == orgUser.OrganizationId))
+        {
+            throw new BadRequestException("You may not join this organization until you leave or remove " +
+                "all other organizations.");
+        }
+
+        // Enforce Single Organization Policy of other organizations user is a member of
+        var anySingleOrgPolicies = await _policyService.AnyPoliciesApplicableToUserAsync(user.Id,
+            PolicyType.SingleOrg);
+        if (anySingleOrgPolicies)
+        {
+            throw new BadRequestException("You cannot join this organization because you are a member of " +
+                "another organization which forbids it");
+        }
+
+        // Enforce Two Factor Authentication Policy of organization user is trying to join
+        if (!await userService.TwoFactorIsEnabledAsync(user))
+        {
+            var invitedTwoFactorPolicies = await _policyService.GetPoliciesApplicableToUserAsync(user.Id,
+                PolicyType.TwoFactorAuthentication, OrganizationUserStatusType.Invited);
+            if (invitedTwoFactorPolicies.Any(p => p.OrganizationId == orgUser.OrganizationId))
+            {
+                throw new BadRequestException("You cannot join this organization until you enable " +
+                    "two-step login on your user account.");
+            }
+        }
+
+        orgUser.Status = OrganizationUserStatusType.Accepted;
+        orgUser.UserId = user.Id;
+        orgUser.Email = null;
+
+        await _organizationUserRepository.ReplaceAsync(orgUser);
+
+        var admins = await _organizationUserRepository.GetManyByMinimumRoleAsync(orgUser.OrganizationId, OrganizationUserType.Admin);
+        var adminEmails = admins.Select(a => a.Email).Distinct().ToList();
+
+        if (adminEmails.Count > 0)
+        {
+            var organization = await _organizationRepository.GetByIdAsync(orgUser.OrganizationId);
+            await _mailService.SendOrganizationAcceptedEmailAsync(organization, user.Email, adminEmails);
+        }
+
+        return orgUser;
+    }
+
     public async Task<OrganizationUser> ConfirmUserAsync(Guid organizationId, Guid organizationUserId, string key,
         Guid confirmingUserId, IUserService userService)
     {
